% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_digit.R
\name{predict_digit}
\alias{predict_digit}
\title{Predicts the digit based on a gray-scale image}
\usage{
predict_digit(image)
}
\arguments{
\item{image}{a 784 (28x28) length numeric vector with gray-scale values between 0-255.}
}
\value{
the predicted digit
}
\description{
Predicts the digit based on a gray-scale image
}
